requirements.txt
Flask==2.3.3
gunicorn==21.2.0
torch==2.2.0
torchvision==0.17.0
pandas==2.1.4
Pillow==10.2.0
numpy==1.26.4
gdown==5.1.0
